requirements.txt
certifi >= 14.05.14
six >= 1.10
python_dateutil >= 2.5.3
setuptools >= 21.0.0
urllib3 >= 1.15.1
python-dotenv >= 0.20.0
pytest >= 7.2.1
pydantic >= 2.5.1
pytest-xdist >= 3.5.0
pytest-randomly >= 3.15.0
boto3 >= 1.34.1
botocore~=1.34.1
pysmb>=1.2.9.1